requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
crispy-bootstrap5==0.7
cssbeautifier==1.14.6
dj-database-url==1.0.0
Django==4.1.2
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
djlint==1.19.1
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
jsbeautifier==1.14.6
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
python-decouple==3.6
PyYAML==6.0
regex==2022.9.13
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.5
zipp==3.9.0
